Factorisation fonctions installation et autoload
