+++ includes/fonction.php
@@ -415,5 +415,3 @@ function deconnexion_CAS($url="")
 
     phpCAS::logoutWithUrl($url);
 }
-
-
