+++ install/upgrade_from_v1.6.0.php
@@ -33,7 +33,6 @@ defined( '_PHP_CONGES' ) or die( 'Restricted access' );
 /*******************************************************************/
 include ROOT_PATH .'fonctions_conges.php' ;
 include INCLUDE_PATH .'fonction.php';
-include 'fonctions_install.php' ;
 
 $PHP_SELF=$_SERVER['PHP_SELF'];
 
@@ -45,20 +44,7 @@ $version = (isset($_GET['version']) ? $_GET['version'] : (isset($_POST['version'
 $lang = (isset($_GET['lang']) ? $_GET['lang'] : (isset($_POST['lang']) ? $_POST['lang'] : "")) ;
 
 //étape 1 création de la table de gestion des plugins
-e1_create_table_plugins();
-
-function e1_create_table_plugins()
-    {
-    $create_table_plugin_query = "CREATE TABLE IF NOT EXISTS `conges_plugins` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `p_name` varchar(250) COLLATE utf8_unicode_ci NOT NULL COMMENT 'Plugin name',
-  `p_is_active` tinyint(1) NOT NULL DEFAULT '0' COMMENT 'Plugin activated ?',
-  `p_is_install` tinyint(1) NOT NULL DEFAULT '0' COMMENT 'Plugin is installed ?',
-  PRIMARY KEY (`id`)
-) ENGINE=MyISAM DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci AUTO_INCREMENT=1 ;";
-    $result_create_table_plugin = \includes\SQL::query($create_table_plugin_query);
-    }
-
+\install\Fonctions::e1_create_table_plugins();
 
     // on renvoit à la page mise_a_jour.php (là d'ou on vient)
     echo "<a href=\"mise_a_jour.php?etape=4&version=$version&lang=$lang\">upgrade_from_v1.6.0  OK</a><br>\n";
